Organize and Simplify ConnectionProfile
This change removes most non-idiomatic implementations
of Go in the connectionprofile package. It strongly types
function input parameters, and removes function calls with
extraneous arguments. It moves variable declarations closer
to the consumers and adds default information to the
configuration structs rather than appending them as bytes.

There is a lot of work to be desired in this package
but this is a start for bring some common sense to the
package.
