+++ tools/operator/fabricconfig/orderer.go
@@ -11,11 +11,9 @@ import (
 	"io/ioutil"
 	"time"
 
-	"gopkg.in/yaml.v2"
-
 	"github.com/hyperledger/fabric-test/tools/operator/networkspec"
 	"github.com/hyperledger/fabric-test/tools/operator/paths"
-	//"github.com/hyperledger/fabric-test/tools/operator/launcher/k8s"
+	"gopkg.in/yaml.v2"
 )
 
 type Orderer struct {
